requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
mysql-connector-python
bcrypt
alembic
starlette